Little bugfixes [50]
Documents/yknytt_data to override working directory
Save visited areas while scrolling in DebugFlyMode (Ctrl+F4)
Do not switch border if 'h' is pressed while typing or outside the game
Show completion status and user score in 'Select Save'
New command 'reboot tilesets' to reload a level without clearing all cache files, refresh only tilesets
Add 'reboot tilesets' to console and bind it to Ctrl+F7
Restore previous high jump speed to make jump height more precise (54px)
Add 2px boost to double jump (this value was obtained empirically, now low double jump = 59px, high double jump = 105px)
Remove LOW_JUMP_EXTRA_GRAVITY
Make IdleState more like WalkRunState (to fix missing double jump when falling from IdleState)
Apply area warp if relative shift goes outside an area
Show down button hint for hidden signs
Fix one of the cases of chain shifts
Fix border of sun_ray_hd.png
Don't accidentally show previous results in 'Download Levels'
Fix lost focus on the search field
Rename 'Unfinished' levels to 'Art/Unfinished'
Do not close console and execute a command in it simultaniously
New 'save cache' command to export cache (with maps)
Do not ask about map tile every time if it's not in a filesystem
Make Juni a little higher (16.1px), useful for 8px platforms

+++ knytt/GDKnyttWorldImpl.cs
@@ -3,6 +3,10 @@ using YKnyttLib;
 
 public class GDKnyttWorldImpl : KnyttWorld
 {
+    public int Completed { get; set; }
+    public int UserScore { get; set; }
+    public bool HasSaves { get; set; }
+
     public GDKnyttWorldImpl() : base() { }
 
     protected override object bytesToSound(byte[] data, bool loop)
@@ -95,7 +99,7 @@ public class GDKnyttWorldImpl : KnyttWorld
         GDKnyttDataStore.ProgressHint = "Unpacking completed.";
     }
 
-    public static void removeDirectory(string dir_name)
+    public static void removeDirectory(string dir_name, string prefix = "")
     {
         var dir = new Directory();
         if (!dir.DirExists(dir_name)) { return; }
@@ -103,6 +107,7 @@ public class GDKnyttWorldImpl : KnyttWorld
         dir.ListDirBegin(skipNavigational: true);
         for (string filename = dir.GetNext(); filename != ""; filename = dir.GetNext())
         {
+            if (!filename.StartsWith(prefix)) { continue; }
             if (dir.FileExists(filename)) { dir.Remove(filename); } else { removeDirectory($"{dir_name}/{filename}"); }
         }
         dir.ListDirEnd();
@@ -122,12 +127,13 @@ public class GDKnyttWorldImpl : KnyttWorld
         removeDirectory(GDKnyttDataStore.BaseDataDirectory.PlusFile("Cache").PlusFile(WorldDirectoryName));
     }
 
-    public void refreshWorld()
+    public void refreshWorld(bool tilesets_only)
     {
         if (BinMode)
         {
             BinLoader = new KnyttBinWorldLoader(GDKnyttAssetManager.loadFile(WorldDirectory));
         }
-        removeDirectory(GDKnyttDataStore.BaseDataDirectory.PlusFile("Cache").PlusFile(WorldDirectoryName));
+        removeDirectory(GDKnyttDataStore.BaseDataDirectory.PlusFile("Cache").PlusFile(WorldDirectoryName), 
+                        prefix: tilesets_only ? "Tileset" : "");
     }
 }
